+++ src/pysubnet/config.py
@@ -14,7 +14,7 @@ class Config:
     root_dir: str = "./network"
     clean: bool = False
     chainspec: str = "dev"
-    raw_chainspec: Path = None # Is generated by us so is a fs path
+    raw_chainspec: Path = None  # Is generated by us so is a fs path
     bin: Path = None
     account_key_type: AccountKeyType = None
     poa: bool = False
@@ -51,6 +51,13 @@ class Config:
 def parse_args() -> Config:
     parser = argparse.ArgumentParser(description="PySubnet launcher")
     parser.add_argument(
+        "--config",
+        "-f",
+        dest="config_file",
+        type=Path,
+        help="Path to the configuration file",
+    )
+    parser.add_argument(
         "-i",
         "--interactive",
         action="store_true",
@@ -82,9 +89,7 @@ def parse_args() -> Config:
         default="dev",
         help="Path to chainspec file or 'dev' | 'local'",
     )
-    parser.add_argument(
-        "--bin", type=Path, help="Path to substrate binary"
-    )
+    parser.add_argument("--bin", type=Path, help="Path to substrate binary")
     parser.add_argument(
         "--account",
         type=AccountKeyType.from_string,
@@ -99,10 +104,8 @@ def parse_args() -> Config:
     )
 
     args = parser.parse_args()
-    # !! WARNING: argsparse will actually set non-supplied flags to None! This works for boolean values but
-    # for others it its a very annoying runtime error! Hence use + <default value> unless a default is provided
-    # in argsparse itself. We explicitly specify defaults for argparse itself so `or <default_val>` not required here
-    return Config(
+
+    config = Config(
         interactive=args.interactive,
         run_network=args.run_network,
         root_dir=os.path.abspath(args.root),
@@ -112,3 +115,10 @@ def parse_args() -> Config:
         account_key_type=args.account,
         poa=args.poa,
     )
+    # !! WARNING: argsparse will actually set non-supplied flags to None! This works for boolean values but
+    # for others it its a very annoying runtime error! Hence use + <default value> unless a default is provided
+    # in argsparse itself. We explicitly specify defaults for argparse itself so `or <default_val>` not required here
+    if args.config_file is not None:
+        nodes = load_nodes_from_file(args.config_file)
+
+    return config
