+++ src/pysubnet/helpers/config_file.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+
+def parse_config_file(config_file: Path):
+    pass
